requirements.txt
google-cloud-speech
google-cloud-storage
srt
ffmpeg-python
